split things up, add config for deployment
