+++ archive/archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/Archivist.scala
@@ -12,7 +12,7 @@ import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.{NotificationMessage, SNSConfig}
 import uk.ac.wellcome.platform.archive.archivist.flow.{BagArchiveCompleteFlow, DownloadZipFileFlow, UploadAndVerifyBagFlow}
 import uk.ac.wellcome.platform.archive.archivist.models.BagUploaderConfig
-import uk.ac.wellcome.platform.archive.common.MessageStream
+import uk.ac.wellcome.platform.archive.common.messaging.MessageStream
 import uk.ac.wellcome.storage.ObjectLocation
 
 import scala.concurrent.ExecutionContext
