+++ archive/common/src/main/scala/uk/ac/wellcome/platform/archive/common/modules/MessageStreamModule.scala
@@ -6,7 +6,7 @@ import com.google.inject.{AbstractModule, Provides}
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.sqs.SQSConfig
 import uk.ac.wellcome.monitoring.MetricsSender
-import uk.ac.wellcome.platform.archive.common.MessageStream
+import uk.ac.wellcome.platform.archive.common.messaging.MessageStream
 
 object MessageStreamModule extends AbstractModule {
   @Provides
