+++ archive/common/src/test/scala/uk/ac/wellcome/platform/archive/common/messaging/MessageStreamTest.scala
@@ -15,7 +15,6 @@ import uk.ac.wellcome.messaging.sqs.SQSConfig
 import uk.ac.wellcome.messaging.test.fixtures.Messaging
 import uk.ac.wellcome.messaging.test.fixtures.SQS.{Queue, QueuePair}
 import uk.ac.wellcome.monitoring.MetricsSender
-import uk.ac.wellcome.platform.archive.common.MessageStream
 import uk.ac.wellcome.platform.archive.common.fixtures.AkkaS3
 import uk.ac.wellcome.storage.utils.ExtendedPatience
 import uk.ac.wellcome.test.fixtures.TestWith
