+++ archive/registrar/src/main/scala/uk/ac/wellcome/platform/archive/registrar/config/ArgsConfigurator.scala
@@ -5,7 +5,8 @@ import uk.ac.wellcome.messaging.sns.SNSConfig
 import uk.ac.wellcome.messaging.sqs.SQSConfig
 import uk.ac.wellcome.monitoring.MetricsConfig
 import uk.ac.wellcome.platform.archive.common.modules._
-import uk.ac.wellcome.platform.archive.registrar.models.{HybridStoreConfig, RegistrarConfig}
+import uk.ac.wellcome.platform.archive.registrar.models.RegistrarConfig
+import uk.ac.wellcome.platform.archive.registrar.modules.HybridStoreConfig
 import uk.ac.wellcome.storage.dynamo.DynamoConfig
 import uk.ac.wellcome.storage.s3.S3Config
