+++ archive/registrar/src/main/scala/uk/ac/wellcome/platform/archive/registrar/models/BagRegistrationCompleteNotification.scala
@@ -1,0 +1,3 @@
+package uk.ac.wellcome.platform.archive.registrar.models
+
+case class BagRegistrationCompleteNotification(storageManifest: StorageManifest)
