+++ archive/registrar/src/main/scala/uk/ac/wellcome/platform/archive/registrar/models/RegistrarConfig.scala
@@ -4,8 +4,7 @@ import uk.ac.wellcome.messaging.sns.SNSConfig
 import uk.ac.wellcome.messaging.sqs.SQSConfig
 import uk.ac.wellcome.monitoring.MetricsConfig
 import uk.ac.wellcome.platform.archive.common.modules._
-import uk.ac.wellcome.storage.dynamo.DynamoConfig
-import uk.ac.wellcome.storage.s3.S3Config
+import uk.ac.wellcome.platform.archive.registrar.modules.HybridStoreConfig
 
 case class RegistrarConfig(s3ClientConfig: S3ClientConfig,
                            cloudwatchClientConfig: CloudwatchClientConfig,
@@ -14,13 +13,6 @@ case class RegistrarConfig(s3ClientConfig: S3ClientConfig,
                            snsClientConfig: SnsClientConfig,
                            snsConfig: SNSConfig,
                            hybridStoreConfig: HybridStoreConfig,
-                           metricsConfig: MetricsConfig)
+                           metricsConfig: MetricsConfig
+                          )
 
-
-case class HybridStoreConfig(
-                              dynamoClientConfig: DynamoClientConfig,
-                              s3ClientConfig: S3ClientConfig,
-                              dynamoConfig: DynamoConfig,
-                              s3Config: S3Config,
-                              s3GlobalPrefix: String
-                            )
